[#28] Abstract main.js
`main.js` was getting *way* too big. I've separated almost all of the functions into their own modules to be imported. This will make it easier when we are ready to support other platforms, as we'll be able to conditionally require modules based on platform.
