+++ modules/common/setupApplicationMenu.js
@@ -1,0 +1,22 @@
+const { Menu } = require('electron')
+
+
+
+
+
+module.exports = function () {
+  Menu.setApplicationMenu(Menu.buildFromTemplate([{
+    label: 'Edit',
+    submenu: [
+      { role: 'undo' },
+      { role: 'redo' },
+      { type: 'separator' },
+      { role: 'cut' },
+      { role: 'copy' },
+      { role: 'paste' },
+      { role: 'pasteandmatchstyle' },
+      { role: 'delete' },
+      { role: 'selectall' },
+    ]
+  }]))
+}
